% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_whisper_path}
\alias{set_whisper_path}
\title{Point to local Whisper Unix Executable File (.exe)}
\usage{
set_whisper_path(path)
}
\arguments{
\item{path}{path to the local Whisper Executable File}
}
\value{
Returns nothing, function sets the option variable
\code{path_to_whisper}.
}
\description{
Function to set an option that points to the local Whisper Executable File
\code{whisper}.
}
\details{
List of possible file path locations for the local Whisper
Executable File
\describe{
\item{Linux}{/usr/bin/whisper, /usr/local/bin/whisper}
\item{Mac}{/opt/homebrew/Caskroom/miniforge/base/bin/whisper}
\item{Windows}{C:\\Program Files\\whisper}
}
}
\examples{
\dontrun{
set_whisper_path("~/your/path/to/whisper")
}
}
