% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate to English}
\usage{
translate(
  audio_file,
  model_name = "small",
  audio_lang,
  output_dir,
  output_format = "txt"
)
}
\arguments{
\item{audio_file}{Input audio file: \code{"m4a"}, \code{"mp3"}, \code{"webm"}, \code{"mp4"}, \code{"mpga"}, \code{"wav"}, or \code{"mpeg"}}

\item{model_name}{Name of the Whisper model. Models include:
\itemize{
\item \code{"tiny.en"}, \code{"tiny"}
\item \code{"base.en"}, \code{"base"}
\item \code{"small.en"}, \code{"small"}
\item \code{"medium.en"}, \code{"medium"}
\item \code{"large-v1"}, \code{"large-v2"}, \code{"large"}
}}

\item{audio_lang}{Language spoken in audio. Specify \code{"None"} to perform language detection.}

\item{output_dir}{Directory to save the output.}

\item{output_format}{Desired format of the output file: \code{"txt"}, \code{"vtt"}, \code{"srt"}, \code{"tsv"}, \code{"json"}, or \code{"all"}.}
}
\value{
A message indicating whether the function was a success or failure
}
\description{
Translate audio in a different language to English.
}
\examples{
\dontrun{
# Korean to English
translate("sample-korean.mp3", audio_lang = "Korean", output_dir = ".")

# Japanese to English
translate("sample-japanese.mp3", audio_lang = "Japanese", output_dir = ".")
}

}
